Move enum AppLock to the Feature extension
