+++ Cartfile.resolved
@@ -3,7 +3,7 @@ github "wireapp/PINCache" "2.3-swift3.1"
 github "wireapp/ocmock" "v3.4.3"
 github "wireapp/swift-protobuf" "1.8.0_Swift5.2.2"
 github "wireapp/wire-ios-cryptobox" "24.0.0"
-github "wireapp/wire-ios-data-model" "feature/save-feature-configs"
+github "wireapp/wire-ios-data-model" "feature/feature-configuration-main"
 github "wireapp/wire-ios-images" "32.0.0"
 github "wireapp/wire-ios-link-preview" "28.0.0"
 github "wireapp/wire-ios-protos" "25.0.0"
